+++ module-check/src/main/clojure/clojure/core/typed/base_env.clj
@@ -841,7 +841,7 @@ clojure.core/promise
 
 clojure.core/deliver (All [x] [(Promise x) x -> (U nil (Promise x))])
 
-clojure.core/flatten [Any -> Any]
+clojure.core/flatten [(U nil (Seqable Any)) -> (Seq Any)]
 
 ;TODO review
 ;clojure.core/group-by (All [x y] [[x -> y] (U nil (Seqable x)) -> (Map y (Vec x))])